Add MeritConsensus
Use MeritConsensus for merit related methods instead of VoteResultConsensus
