+++ src/main/java/bisq/core/dao/governance/voteresult/VoteResultConsensus.java
@@ -18,25 +18,17 @@
 package bisq.core.dao.governance.voteresult;
 
 import bisq.core.dao.governance.blindvote.VoteWithProposalTxIdList;
-import bisq.core.dao.governance.merit.MeritList;
 import bisq.core.dao.state.BsqStateService;
 import bisq.core.dao.state.blockchain.Tx;
 import bisq.core.dao.state.blockchain.TxInput;
 import bisq.core.dao.state.blockchain.TxOutput;
 import bisq.core.dao.state.blockchain.TxOutputType;
 import bisq.core.dao.state.blockchain.TxType;
-import bisq.core.dao.state.governance.Issuance;
 import bisq.core.dao.state.period.DaoPhase;
 import bisq.core.dao.state.period.PeriodService;
 
 import bisq.common.crypto.Encryption;
-import bisq.common.util.MathUtils;
 import bisq.common.util.Utilities;
-
-import org.bitcoinj.core.ECKey;
-import org.bitcoinj.core.Sha256Hash;
-
-import com.google.common.annotations.VisibleForTesting;
 
 import javax.crypto.SecretKey;
 
@@ -53,8 +45,6 @@ import static com.google.common.base.Preconditions.checkArgument;
 
 @Slf4j
 public class VoteResultConsensus {
-    private static final int BLOCKS_PER_YEAR = 50_000; // 51264;
-
     public static boolean hasOpReturnDataValidLength(byte[] opReturnData) {
         return opReturnData.length == 38;
     }
@@ -71,126 +61,6 @@ public class VoteResultConsensus {
         } catch (Throwable t) {
             throw new VoteResultException(t);
         }
-    }
-
-    public static MeritList decryptMeritList(byte[] encryptedMeritList, SecretKey secretKey) throws VoteResultException {
-        try {
-            final byte[] decrypted = Encryption.decrypt(encryptedMeritList, secretKey);
-            return MeritList.getMeritListFromBytes(decrypted);
-        } catch (Throwable t) {
-            throw new VoteResultException(t);
-        }
-    }
-
-    // TODO move ot MeritConsensus
-    public static long getWeightedMeritAmount(long amount, int issuanceHeight, int blockHeight, int blocksPerYear) {
-        if (issuanceHeight > blockHeight)
-            throw new IllegalArgumentException("issuanceHeight must not be larger than blockHeight");
-        if (blockHeight < 0)
-            throw new IllegalArgumentException("blockHeight must not be negative");
-        if (amount < 0)
-            throw new IllegalArgumentException("amount must not be negative");
-        if (blocksPerYear < 0)
-            throw new IllegalArgumentException("blocksPerYear must not be negative");
-
-        // We use a linear function  to apply a factor for the issuance amount of 1 if the issuance was recent and 0
-        // if the issuance was 2 years old or older.
-
-        // TODO Use stairway function without double/divisions or lookup table
-        int maxAge = 2 * blocksPerYear;
-        int age = Math.min(maxAge, blockHeight - issuanceHeight);
-
-        double rel = MathUtils.roundDouble((double) age / (double) maxAge, 10);
-        double factor = 1d - rel;
-        long weightedAmount = MathUtils.roundDoubleToLong(amount * factor);
-        log.debug("getWeightedMeritAmount: amount={}, factor={}, weightedAmount={}, ", amount, factor, weightedAmount);
-        return weightedAmount;
-    }
-
-    public static long getMeritStake(String blindVoteTxId, MeritList meritList, BsqStateService bsqStateService) {
-        int txChainHeight = bsqStateService.getTx(blindVoteTxId).map(Tx::getBlockHeight).orElse(0);
-        return getMeritStake(blindVoteTxId, meritList, txChainHeight);
-    }
-
-    @VisibleForTesting
-    static long getMeritStake(String blindVoteTxId, MeritList meritList, int txChainHeight) {
-        // We need to take the chain height when the blindVoteTx got published so we get the same merit for the vote even at
-        // later blocks (merit decreases with each block).
-        if (txChainHeight == 0) {
-            log.error("Error at getMeritStake: blindVoteTx not found in bsqStateService. blindVoteTxId=" + blindVoteTxId);
-            return 0;
-        }
-
-        return meritList.getList().stream()
-                .filter(merit -> {
-                    // TODO make method
-                    // We verify if signature of hash of blindVoteTxId is correct. EC key from first input for blind vote tx is
-                    // used for signature.
-                    String pubKeyAsHex = merit.getIssuance().getPubKey();
-                    if (pubKeyAsHex == null) {
-                        log.error("Error at getMeritStake: pubKeyAsHex is null");
-                        return false;
-                    }
-
-                    // TODO Check if a sig key was used multiple times for different voters
-                    // At the moment we don't impl. that to not add too much complexity and as we consider that
-                    // risk very low.
-
-                    boolean result = false;
-                    try {
-                        ECKey pubKey = ECKey.fromPublicOnly(Utilities.decodeFromHex(pubKeyAsHex));
-                        ECKey.ECDSASignature signature = ECKey.ECDSASignature.decodeFromDER(merit.getSignature()).toCanonicalised();
-                        Sha256Hash msg = Sha256Hash.wrap(blindVoteTxId);
-                        result = pubKey.verify(msg, signature);
-                    } catch (Throwable t) {
-                        log.error("Signature verification of issuance failed: " + t.toString());
-                    }
-                    if (!result) {
-                        log.error("Signature verification of issuance failed: blindVoteTxId={}, pubKeyAsHex={}",
-                                blindVoteTxId, pubKeyAsHex);
-                    }
-                    return result;
-                })
-                .mapToLong(merit -> {
-                    try {
-                        int issuanceHeight = merit.getIssuance().getChainHeight();
-                        if (issuanceHeight <= txChainHeight) {
-                            return getWeightedMeritAmount(merit.getIssuance().getAmount(),
-                                    issuanceHeight,
-                                    txChainHeight,
-                                    BLOCKS_PER_YEAR);
-                        } else {
-                            return 0;
-                        }
-                    } catch (Throwable t) {
-                        log.error("Error at getMeritStake: " + t.toString());
-                        return 0;
-                    }
-                })
-                .sum();
-    }
-
-    // Used to get the currently available merit before we have made the blind vote.
-    public static long getCurrentlyAvailableMerit(MeritList meritList, int currentChainHeight) {
-        // We need to take the chain height when the blindVoteTx got published so we get the same merit for the vote even at
-        // later blocks (merit decreases with each block).
-        return meritList.getList().stream()
-                .mapToLong(merit -> {
-                    try {
-                        Issuance issuance = merit.getIssuance();
-                        int issuanceHeight = issuance.getChainHeight();
-                        checkArgument(issuanceHeight <= currentChainHeight,
-                                "issuanceHeight must not be larger as currentChainHeight");
-                        return getWeightedMeritAmount(issuance.getAmount(),
-                                issuanceHeight,
-                                currentChainHeight,
-                                BLOCKS_PER_YEAR);
-                    } catch (Throwable t) {
-                        log.error("Error at getCurrentlyAvailableMerit: " + t.toString());
-                        return 0;
-                    }
-                })
-                .sum();
     }
 
     // We compare first by stake and in case we have multiple entries with same stake we use the
